requirements.txt
annotated-types==0.6.0
antiorm==1.2.1
anyio==3.7.1
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.2
certifi==2023.11.17
click==8.1.7
db==0.1.1
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
loguru==0.7.2
outcome==1.3.0.post0
playwright==1.40.0
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyee==11.0.1
PySocks==1.7.1
python-dotenv==1.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
wsproto==1.2.0
